Add amend commit functionality to CI script
This commit introduces the ability to amend the last commit in the CI
script. It adds an argument parser to handle the '--amend' flag. If the
flag is present, the script will amend the last commit, otherwise it
will create a new one as before. The commit message is generated based
on the git diff.