Change default history value to 0 in ci command
The default value for the `--history` option in the `ci` command has been
changed from 1 to 0. This means that by default, no history will be included
in the commit message prompt.

The commit message creation logic has been updated to handle a history
value of 0 by setting the history messages to an empty list.